+++ commons/include/elf_section_header_builders.h
@@ -14,10 +14,10 @@
 //——————————————————————————————————————————————————————————————————————————————
 
 lang_status_t shstrtab_init    (buffer_t* tab, size_t init_capacity);
-lang_status_t add_section_name (buffer_t* tab, const char* name, size_t* offset);
+lang_status_t add_section_name (buffer_t* tab, const char* name, uint32_t* offset);
 
-void build_text_shdr(Elf64_Shdr* shdr, size_t code_size, size_t name_offset);
-void build_shstrtab_shdr(Elf64_Shdr* shdr, size_t shstrtab_size, size_t name_offset, size_t table_offset);
+void build_text_shdr(Elf64_Shdr* shdr, size_t code_size, uint32_t name_offset);
+void build_shstrtab_shdr(Elf64_Shdr* shdr, size_t shstrtab_size, uint32_t name_offset, size_t table_offset);
 
 //——————————————————————————————————————————————————————————————————————————————
